Make small changes
