+++ src/extraction/tablenet/__init__.py
@@ -2,5 +2,6 @@
 
 from .marmot import MarmotDataModule
 from .tablenet import TableNetModule, DiceLoss
+from .metrics import binary_mean_iou
 
-__all__ = ["MarmotDataModule", "TableNetModule", "DiceLoss"]
+__all__ = ["MarmotDataModule", "TableNetModule", "DiceLoss", "binary_mean_iou"]
