+++ src/extraction/tablenet/tablenet.py
@@ -10,8 +10,7 @@ from PIL import Image
 from torch import nn, optim
 from torch.nn import functional as F
 from torchvision.models import vgg19, vgg19_bn, VGG19_Weights, VGG19_BN_Weights
-
-EPSILON = 1e-15
+from .metrics import binary_mean_iou
 
 
 class TableNetModule(pl.LightningModule):
@@ -194,39 +193,43 @@ class TableNetModule(pl.LightningModule):
             output_table (torch.Tensor): Batch of table model outputs.
             output_column (torch.Tensor): Batch of column model outputs.
         """
-        # TODO: clean up
         for i in range(min(len(labels_table), n_images)):
             image = Image.fromarray(
                 255 * labels_table[i].squeeze().cpu().numpy().astype(np.uint8)
             )
             file_name = f"table_label_{i}.png"
-            image.save(file_name)
-            mlflow.log_artifact(file_name, artifact_path="images")
-            os.remove(file_name)
+            self._log_image(image, file_name)
 
             image = Image.fromarray(
                 255 * labels_column[i].squeeze().cpu().numpy().astype(np.uint8)
             )
             file_name = f"column_label_{i}.png"
-            image.save(file_name)
-            mlflow.log_artifact(file_name, artifact_path="images")
-            os.remove(file_name)
+            self._log_image(image, file_name)
 
             image = Image.fromarray(
                 255 * output_table[i].squeeze().cpu().numpy().astype(np.uint8)
             )
             file_name = f"table_output_{i}.png"
-            image.save(file_name)
-            mlflow.log_artifact(file_name, artifact_path="images")
-            os.remove(file_name)
+            self._log_image(image, file_name)
 
             image = Image.fromarray(
                 255 * output_column[i].squeeze().cpu().numpy().astype(np.uint8)
             )
             file_name = f"column_output_{i}.png"
-            image.save(file_name)
-            mlflow.log_artifact(file_name, artifact_path="images")
-            os.remove(file_name)
+            self._log_image(image, file_name)
+
+    @staticmethod
+    def _log_image(image, file_name):
+        """
+        Log a single image with the given file name.
+
+        Args:
+            image (Image): Image.
+            file_name (str): File name.
+        """
+        image.save(file_name)
+        mlflow.log_artifact(file_name, artifact_path="images")
+        os.remove(file_name)
 
 
 class TableNet(nn.Module):
@@ -251,7 +254,7 @@ class TableNet(nn.Module):
         )
         for param in self.vgg.parameters():
             param.requires_grad = False
-        self.layers = [18, 27] if not batch_norm else [26, 39]
+        self.feature_maps_ids = [18, 27] if not batch_norm else [26, 39]
         self.table_decoder = TableDecoder(num_class)
         self.column_decoder = ColumnDecoder(num_class)
 
@@ -262,21 +265,20 @@ class TableNet(nn.Module):
         Args:
             x (tensor): Batch of images to perform forward-pass.
 
-        Returns (Tuple[tensor, tensor]): Table, Column prediction.
+        Returns (Tuple[torch.Tensor, torch.Tensor]): Table, Column prediction.
         """
-        results = []
-        for i, layer in enumerate(self.vgg):
-            x = layer(x)
-            if i in self.layers:
-                results.append(x)
+        feature_maps = []
+        with torch.no_grad():
+            for i, layer in enumerate(self.vgg):
+                x = layer(x)
+                if i in self.feature_maps_ids:
+                    feature_maps.append(x)
 
-        x_table = self.table_decoder(x, results)
+        x_table = self.table_decoder(x, feature_maps)
         table_output = torch.sigmoid(x_table)
-        del x_table
 
-        x_column = self.column_decoder(x, results)
+        x_column = self.column_decoder(x, feature_maps)
         column_output = torch.sigmoid(x_column)
-        del x_column
         return table_output, column_output
 
 
@@ -378,22 +380,3 @@ class DiceLoss(nn.Module):
         )
 
         return 1 - dice
-
-
-def binary_mean_iou(inputs, targets):
-    """Calculate binary mean intersection over union.
-
-    Args:
-        inputs (tensor): Output from the forward pass.
-        targets (tensor): Labels.
-
-    Returns (tensor): Intersection over union value.
-    """
-    output = (inputs > 0).int()
-    if output.shape != targets.shape:
-        targets = torch.squeeze(targets, 1)
-    intersection = (targets * output).sum()
-    union = targets.sum() + output.sum() - intersection
-    result = (intersection + EPSILON) / (union + EPSILON)
-
-    return result
